+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
+### Added
+- Load process log into flowchart
+
+### Changed
+- Moved ProcessLogger to ProcessingPlugin namespace
+- Removed get methods from ProcessLogger interface
+- Renamed table processes to link_pm_process_log
+- Removed outdated ProcessViewController
 
 ## [0.3.2] - 2015-05-12
 ### Fixed